convert to controllerAs for employeeInformation

+++ ums-api/src/main/java/org/ums/domain/model/immutable/registrar/AcademicInformation.java
@@ -8,7 +8,7 @@ import org.ums.enums.common.AcademicDegreeType;
 
 import java.io.Serializable;
 
-public interface AcademicInformation extends Serializable, EditType<MutableAcademicInformation>, Identifier<Integer>,
+public interface AcademicInformation extends Serializable, EditType<MutableAcademicInformation>, Identifier<Long>,
     LastModifier {
 
   String getEmployeeId();
